# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Server for uniswap transactions operations.",
        "title": "Uniswap Transaction Tracker",
        "contact": {
            "name": "Alvin Chee",
            "url": "https://github.com/CodingCookieRookie",
            "email": "alvinchee98@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/swap": {
            "get": {
                "description": "returns uniswap swap price with corresponding transaction hash",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "accounts"
                ],
                "summary": "returns uniswap swap price with corresponding transaction hash",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Transaction hash",
                        "name": "txn_hash",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.SwapResponse"
                        }
                    },
                    "500": {
                        "description": "Server Error",
                        "schema": {
                            "$ref": "#/definitions/errors.ServerError"
                        }
                    }
                }
            }
        },
        "/transaction/fee": {
            "get": {
                "description": "returns transaction fee of transaction with corresponding transaction hash",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "accounts"
                ],
                "summary": "returns transaction fee of transaction with corresponding transaction hash",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Transaction hash",
                        "name": "txn_hash",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.TxnFeeResp"
                        }
                    },
                    "500": {
                        "description": "Server Error",
                        "schema": {
                            "$ref": "#/definitions/errors.ServerError"
                        }
                    }
                }
            }
        },
        "/transactions": {
            "get": {
                "description": "returns all historical transactions from start to end time",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "accounts"
                ],
                "summary": "returns all historical transactions from start to end time",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Start Time in ISO 8601 format",
                        "name": "start_time",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "End Time in ISO 8601 format",
                        "name": "end_time",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.TxnsResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.UserError"
                        }
                    },
                    "500": {
                        "description": "Server Error",
                        "schema": {
                            "$ref": "#/definitions/errors.ServerError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "errors.ServerError": {
            "type": "object",
            "properties": {
                "msg": {
                    "type": "string"
                }
            }
        },
        "errors.UserError": {
            "type": "object",
            "properties": {
                "msg": {
                    "type": "string"
                }
            }
        },
        "model.SwapEventWithToken": {
            "type": "object",
            "properties": {
                "from": {
                    "type": "string"
                },
                "sqrtPriceX96": {
                    "type": "string"
                },
                "to": {
                    "type": "string"
                },
                "tokenDecimal": {
                    "type": "string"
                },
                "tokenName": {
                    "type": "string"
                },
                "tokenSymbol": {
                    "type": "string"
                }
            }
        },
        "model.SwapPrice": {
            "type": "object",
            "properties": {
                "price": {
                    "type": "string"
                },
                "swap_event": {
                    "$ref": "#/definitions/model.SwapEventWithToken"
                }
            }
        },
        "model.SwapResponse": {
            "type": "object",
            "properties": {
                "swap_prices": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.SwapPrice"
                    }
                }
            }
        },
        "model.Txn": {
            "type": "object",
            "properties": {
                "blockHash": {
                    "type": "string"
                },
                "blockNumber": {
                    "type": "string"
                },
                "confirmations": {
                    "type": "string"
                },
                "contractAddress": {
                    "type": "string"
                },
                "cumulativeGasUsed": {
                    "type": "string"
                },
                "ethprice": {
                    "type": "number"
                },
                "from": {
                    "type": "string"
                },
                "gas": {
                    "type": "string"
                },
                "gasPrice": {
                    "type": "string"
                },
                "gasUsed": {
                    "type": "string"
                },
                "hash": {
                    "type": "string"
                },
                "input": {
                    "type": "string"
                },
                "nonce": {
                    "type": "string"
                },
                "timeStamp": {
                    "type": "string"
                },
                "to": {
                    "type": "string"
                },
                "tokenDecimal": {
                    "type": "string"
                },
                "tokenName": {
                    "type": "string"
                },
                "tokenSymbol": {
                    "type": "string"
                },
                "transactionIndex": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "model.TxnFeeResp": {
            "type": "object",
            "properties": {
                "transactionFee": {
                    "type": "number"
                }
            }
        },
        "model.TxnsResp": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "result": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Txn"
                    }
                },
                "status": {
                    "type": "string"
                }
            }
        }
    }
}